Added simple code to determine kernel versions
This is a rather simple code to determine the versions of an kernel
installed in the current running system. This code uses the `dnf` api
for the Fedora, CentOS Linux, RHEL and such. This code is also
compatible with the python2.7, as well, python3.8+. The only requirement
is that the user have installed the `python3-dnf` pacakge for python3+,
or, `python2-dnf` package for python2.